p4r_profile_generation/requirements.txt
pandas>=2.0.0
groq>=0.4.0
python-dotenv>=1.0.0